requirements.txt
# Dependencies for Personal AI Employee - Bronze Tier
watchdog>=4.0.0
playwright>=1.40.0
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
google-auth>=2.23.0
# Dependencies for Silver Tier
pydantic>=2.0.0
schedule>=1.2.0
APScheduler>=3.10.0
aiohttp>=3.13.3